- Started adding patch meshes - Fixed revision tracking system for options, attribs and transforms -- Revision count's pointer is now shared so that the revision never goes back because of a stack pop

+++ RI_System/src/RI_State_Error.cpp
@@ -132,7 +132,7 @@ bool State::verifyBasis( RtToken basis, int steps )
 		RI_POWERBASIS,		4,
 	};
 
-	for (int i=0; i < _countof(table); ++i)
+	for (int i=0; i < (int)_countof(table); ++i)
 	{
 		if ( table[i].token == basis )
 		{
